feat: 약관동의 순서 변경
